! Alert'ы внутри VM + ToolBar заменен картинкой

+++ app/src/main/java/by/wlad/koshelev/apexlegendstracker/Arch/VM.kt
@@ -1,11 +1,17 @@
 package by.wlad.koshelev.apexlegendstracker.Arch
 
+import android.app.AlertDialog
 import android.app.Application
+import android.content.DialogInterface
+import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.MutableLiveData
 import by.wlad.koshelev.apexlegendstracker.GamerStats._GamerStats
+import by.wlad.koshelev.apexlegendstracker.R
 import kotlinx.android.synthetic.main.fragment_search.*
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.MainScope
+import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 
 
@@ -106,10 +112,26 @@ class VM(private val apl: Application) : AndroidViewModel(apl) {
      * удаление ОДНОГО игрока
      */
     // удаляем игрока. обновляем список всех. делаем проверку.
-    suspend fun deleteOneGamer(gamer: _GamerStats) = withContext(Dispatchers.IO) {
-        localModel.deleteOneGamer(gamer)
-        getAllGamers()
-        checkLocalGms()
+    fun deleteOneGamer(gamer: _GamerStats, app: AppCompatActivity) {
+        AlertDialog.Builder(app)
+            .setCancelable(false)
+            .setTitle(app.getString(R.string.suuure))
+            .setMessage("${app.getString(R.string.deleteOneGamer)} ${gamer.data.platformInfo.platformUserId} ?")
+
+            .setPositiveButton(app.getString(R.string.yes), DialogInterface.OnClickListener { dialog, which ->
+                MainScope().launch {
+                    localModel.deleteOneGamer(gamer)
+                    getAllGamers()
+                    checkLocalGms()
+                }
+
+            })
+
+            .setNegativeButton(app.getString(R.string.no), DialogInterface.OnClickListener { dialog, which ->
+                dialog.cancel()
+            })
+            .create()
+            .show()
     }
 
 
@@ -117,17 +139,33 @@ class VM(private val apl: Application) : AndroidViewModel(apl) {
      * удаление ВСЕХ из бд.
      */
     // удаляем всех. обновляем всех. проверка.
-    suspend fun dellAllGamers() = withContext(Dispatchers.IO) {
-        localModel.dellAllGamers()
-        getAllGamers()
-        checkLocalGms()
+    fun dellAllGamers(app: AppCompatActivity) {
+        AlertDialog.Builder(app)
+            .setCancelable(false)
+            .setTitle(apl.getString(R.string.suuure))
+            .setMessage(apl.getString(R.string.deleteAllGamers))
+
+            .setPositiveButton(apl.getString(R.string.yes), DialogInterface.OnClickListener { dialog, which ->
+                MainScope().launch {
+                    localModel.dellAllGamers()
+                    getAllGamers()
+                    checkLocalGms()
+                }
+            })
+
+            .setNegativeButton(apl.getString(R.string.no), DialogInterface.OnClickListener { dialog, which ->
+                dialog.cancel()
+            })
+            .create()
+            .show()
+
     }
 
 
     /**
      * локальная проверка
      */
-    // этот метод нужно в методах удаления.
+    // этот метод нужен  в методах удаления.
     // если мы удалим ТЕКУЩЕГО игрока, то в окне подробной статы останется "дата последенго сохранения".
     // этот метод исправляет этот баг
     private suspend fun checkLocalGms() = withContext(Dispatchers.Main) {
